+++ src/main/java/mp_pprl/incremental_clustering/SimilarityCalculator.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 public class SimilarityCalculator {
     /*This method uses the secure summation protocol to create counting bloom filters and then calculates the metric.*/
-    public static double averageSimilaritySecure(Cluster cluster, BloomFilterEncodedRecord bloomFilterEncodedRecord, int bloomFilterLength) {
+    public static float averageSimilaritySecure(Cluster cluster, BloomFilterEncodedRecord bloomFilterEncodedRecord, int bloomFilterLength) {
         List<CountingBloomFilter> countingBloomFilterList = new ArrayList<>();
         for (BloomFilterEncodedRecord clusteredBloomFilterEncodedRecord : cluster.bloomFilterEncodedRecordsSet()) {
             List<BloomFilterEncodedRecord> bloomFilterEncodedRecordsForSummation = new ArrayList<>();
@@ -20,7 +20,7 @@ public class SimilarityCalculator {
             CountingBloomFilter cbf = SummationProtocol.execute(bloomFilterEncodedRecordsForSummation, bloomFilterLength);
             countingBloomFilterList.add(cbf);
         }
-        double sumSimilarity = 0;
+        float sumSimilarity = 0;
         for (CountingBloomFilter cbf : countingBloomFilterList) {
             sumSimilarity += calculateDiceCoefficient(cbf);
         }
@@ -28,8 +28,8 @@ public class SimilarityCalculator {
         return sumSimilarity / countingBloomFilterList.size();
     }
 
-    public static double averageSimilarity(Cluster cluster, BloomFilterEncodedRecord bloomFilterEncodedRecord) {
-        double sumSimilarity = 0;
+    public static float averageSimilarity(Cluster cluster, BloomFilterEncodedRecord bloomFilterEncodedRecord) {
+        float sumSimilarity = 0;
         for (BloomFilterEncodedRecord clusteredRecord : cluster.bloomFilterEncodedRecordsSet()) {
 
             sumSimilarity += calculateSimilarity(bloomFilterEncodedRecord, clusteredRecord);
@@ -38,7 +38,7 @@ public class SimilarityCalculator {
         return sumSimilarity / cluster.bloomFilterEncodedRecordsSet().size();
     }
 
-    private static double calculateDiceCoefficient(CountingBloomFilter countingBloomFilter) {
+    private static float calculateDiceCoefficient(CountingBloomFilter countingBloomFilter) {
         int[] vector = countingBloomFilter.getVector();
 
         int numberOfMatches = 0;
@@ -50,10 +50,10 @@ public class SimilarityCalculator {
             cbfSum += counter;
         }
 
-        return (double) (countingBloomFilter.getNumberOfBloomFilters() * numberOfMatches) / cbfSum;
+        return (float) (countingBloomFilter.getNumberOfBloomFilters() * numberOfMatches) / cbfSum;
     }
 
-    private static double calculateSimilarity(BloomFilterEncodedRecord r1, BloomFilterEncodedRecord r2) {
+    private static float calculateSimilarity(BloomFilterEncodedRecord r1, BloomFilterEncodedRecord r2) {
         byte[] bf1 = r1.getBloomFilter().getVector();
         byte[] bf2 = r2.getBloomFilter().getVector();
         int[] cbf = new int[bf1.length];
@@ -71,6 +71,6 @@ public class SimilarityCalculator {
             }
         }
 
-        return (double) (2 * numberOfMatches) / cbfSum;
+        return (float) (2 * numberOfMatches) / cbfSum;
     }
 }
